Added randomness
Now, when the user chooses an option, the next event is randomly chosen in a pool of possible events. Thus, the life of a character is not only determined by the choices of the user, but also by randomness. There can be up to 3 possible events to randomly choose from.